add type info
