+++ core/amber/src/main/python/core/architecture/managers/console_message_manager.py
@@ -1,4 +1,4 @@
-from core.util.console_message.timed_buffer import TimedBuffer
+from core.util.buffer.timed_buffer import TimedBuffer
 
 
 class ConsoleMessageManager:
