+++ core/amber/src/main/python/core/util/buffer/buffer_base.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta
+
+from core.util.protocol.base_protocols import FlushedGetable, Putable
+
+
+class IBuffer(FlushedGetable, Putable, metaclass=ABCMeta):
+    pass
