+++ core/amber/src/main/python/core/util/protocol/base_protocols.py
@@ -1,0 +1,41 @@
+from abc import abstractmethod
+from typing import Protocol, TypeVar, Sized, Optional
+
+T = TypeVar("T")
+K = TypeVar("K")
+
+
+class Putable(Protocol):
+    @abstractmethod
+    def put(self, item: T) -> None:
+        pass
+
+
+class KeyedPutable(Protocol):
+    @abstractmethod
+    def put(self, key: K, item: T) -> None:
+        pass
+
+
+class Getable(Protocol):
+    @abstractmethod
+    def get(self) -> T:
+        pass
+
+
+class FlushedGetable(Protocol):
+    @abstractmethod
+    def get(self, flush: bool) -> T:
+        pass
+
+
+class EmtpyCheckable(Sized):
+    @abstractmethod
+    def is_empty(self) -> bool:
+        pass
+
+
+class KeyedEmtpyCheckable(Sized):
+    @abstractmethod
+    def is_empty(self, key: Optional[K] = None) -> bool:
+        pass
